+++ tests/execution_spec.rs
@@ -895,12 +895,14 @@ async fn assert_spec(spec: ExecutionSpec, opentelemetry: &InMemoryTelemetry) {
                     identity,
                     tailcall_prettier::Parser::detect(path_str.as_str()).unwrap(),
                 )
+                .await
                 .unwrap();
 
                 let content = tailcall_prettier::format(
                     content,
                     tailcall_prettier::Parser::detect(path_str.as_str()).unwrap(),
                 )
+                .await
                 .unwrap();
 
                 pretty_assertions::assert_eq!(
